refactor(rolldown_binding): move js plugin filter logic to js_plugin_filter.rs
